+++ R/Extract.R
@@ -27,7 +27,7 @@
 #'
 #' @examples
 #' ## Statistics on life forms
-#' summary(as.factor(Easplist$lf_behn_2018))
+#' summary(as.factor(Easplist$life_form))
 #'
 #' ## First ten concepts in this list
 #' summary(Easplist[1:10, ], "all")