+++ R/taxon_traits.R
@@ -38,7 +38,7 @@
 #' # Update
 #' Easplist <- update_trait(
 #'   taxlist = Easplist, ConceptID = 355,
-#'   lf_behn_2018 = "annual"
+#'   life_form = "annual"
 #' )
 #' accepted_name(taxlist = Easplist, ConceptID = 355, show_traits = TRUE)
 #' @rdname taxon_traits